Add initial code lens support
The code lenses will indicate methods that are overrides or interface 
implementations as well as property overrides. This will replace the 
gutter annotations the Atom package currently manages by itself to 
display this information.

The language server protocol has no way to support gutter annotations 
(although in theory, the client could opt to use gutter annotations to 
support code lenses).

There is still more work to be done here: I've already done some 
prototyping for a client implementation in the Atom client, but the 
server will also need support for "workspace/executeCommand" in order to 
actually execute the command in the code lens once the user clicks it in 
the client, and finally will need some way to ask the client to open a 
URI at a specific location in order to navigate to the element being 
overridden or implemented.

In any case, we'll need command execution sooner or later anyway, as it 
is also required for implementing refactoring actions and other commands 
later on.

References #93