Make (first) property range consistent with method range
Property ranges previously pointed to the property name only, whilst 
method ranges include everything: the access modifier, the optional 
static keyword, and so on. This is inconsistent and also causes code 
lenses for properties to be given an indented range, which is also 
inconsistent with method code lenses.

To avoid overlapping ranges, optional second properties in the same 
property block (in case anyone uses those) still only include the 
property.

References #93