Fix unexpected completion sorting by clients due to string sort
Using a number for the sort text to force clients to maintain the same 
order as we generate was a good idea, but unfortunately this number is 
inside a string, which means string sorting rules apply. This means that 
e.g. "2" comes after "19" instead of before it (because "2" comes after 
"1" of "19", even though 2 comes before 19 as a number).

This fixes the issue by padding the numbers with zeroes, so they are of 
the same length.