Attempt to fix code coverage diffs not being displayed for fork MRs
When a fork is created, the code coverage pattern from the project 
settings is not copied over. As pipelines for MRs from forks trigger on 
the _fork_ repository, no code coverage is parsed there - unless the 
user explicitly configured it in the settings screen. This means we lose 
code coverage diffs for these MRs.

Apparently you can also set this pattern inside the CI configuration, 
which is automatically copied over to forks, so this should fix the 
issue without inconveniencing contributors with forks.