Try using workspace folders received durinig initialization
This happens only if no explicit configuration is passed to support 
multi-folder setups without forcing explicit configuration and it will 
still fall back to the rootUri if the client does not support passing 
workspace folders (i.e. it is set to null).

Clients that do support it will very likely also set the 
workspaceFolders to contain the folder of the rootUri in single-project 
setups without explicit configuration.