Always return array for diagnostic relatedInformation
This fix makes sure that an array is always returned for the relatedInformation in diagnostics. The specs indicate that `relatedInformation` should either be omitted or return an array. Returning null caused issues with for example Sublime-LSP (Sublime Text 3).

Imho conditionally omitting the `relatedInformation` is too much hassle so I chose to return an empty array on null instead.