Force file contents to be refetched after save without contents
This forces the contents to be refetched from disk. isset('') returns
true, so using an empty string would cause the next fetch of its
contents to return an empty string, until the next time the contents
were passed.

With this change, a didSave without "text" is treated as the client
saying "the file changed, but I don't have the contents", so the
server will assume that the contents have changed somehow
and persisted to disk, refetching them when necessary.