Fix most PHPStan level 3 issues
The remainder of issues are due to the Promise dependency not correctly
narrowing down the type of the implementation via a docblock or return
type. I've sent a pull request for this upstream [1]. We could also fix
this ourselves, but then we would need to manually override all these
variables to indicate it's a Promise, which is cumbersome.

[1] https://github.com/reactphp/promise/pull/156

References #277