Don't return error response on "File not found" errors
I'm not particularly happy with this solution. If a client requests
functionality in a document the server does not know, an error
response sounds appropriate because the latter cannot sanely handle
the request. Returning empty responses misleads the client into
thinking sort of processing happened and there were no results.
Yet, VSCode insists on throwing response errors in the user's face, so
this sends a warning notification instead.

References #307