Don't fail handling requests when duplicate imports exist
We are currently sending back error responses to requests in several
functionalities when this happens. This is not even really a problem for
us directly, but mostly PhpParser that throws an error in this case.

This fixes some cases that were not using a collecting handler, but a
(default) throwing handler instead. Thanks to this fix, some
requests can now still be appropriately handled in these situations.

References #307