Log warning instead of returing error response when position is unknown
The name qualification utilities we use throw exceptions when you
request information about a position that is not part of the file or
otherwise unknown.

This sounds sane, but the same problem as with unknown files
happens for clients such as VSCode, which throw the error in the face
of the user via the bottom panel, whilst they can do nothing to fix it.

This logs a warning (via a notification from the server) instead.

Fixes #307