Add remaining PHPStan exception problems to baseline for now
We want to fix these at some point in the future, but we don't want CI
to fail for the moment. A load of PHPStan issues have already been fixed
in the last weeks, but there are still a couple hundred remaining.

This way, new code can also already adhere to the appropriate quality
standards and we can fix legacy code steadily.

References #277