Compile the dependency container to allow autowiring in the future
This will also remove unused dependencies (which we currently don't
have, since everything is public).

It is now possible to use basic autowiring. Currently this is still a
pain, since autowiring will look for the class names and we've manually
named everything. Renaming the services to use the class name directly
will work, or aliassing will also work.

Unfortunately, autowiring still only seems to go one level deep:
omitting parameters will look for services with that class as name
automatically, but it will not go further to create those dependencies
automatically as well yet. I'm not sure if that is possible outside
Symfony applications.