Fix type deduction for classlikes followed by generic syntax
Generics are not actually resolved yet, but this fixes not getting
autocompletion or proper responses for other functionality for types such
as Vector<int,string>. For now, Serenata will pretend this just says
"Vector" instead of breaking completely because it can't resolve generics
yet. This is not entirely correct, of course, but better than nothing until
we support generics properly.

References #311