Fix installation not working with Composer 2.0
doctrine/orm requires ocramius/package-versions, but it isn't compatible
with Composer 2 until the latest release (1.8?), which requires PHP 7.4.
We don't want to drop PHP 7.1, 7.2 and 7.3 yet, so it is recommended to
replace it with composer/package-versions-deprecated, which is
compatible all the way back to PHP 7.0.

This fixes installing with Composer 2.0 for those who are updating early,
want to test it, or use it for other reasons.

NOTE: We don't actually require this package ourselves, so this should
probably be removed again at some point in the future.