Test Composer 2 in CI
This tests the upcoming Composer 2, which we should be compatible with
already. This should also improve performance a bit.

This also switches to a specific Composer version so we can verify its
checksum during CI, which should prevent contaminated binaries from
ending up in our CI, in the event that this ever happens.