Fix insertText not having reserved snippet characters escaped properly
This is only relevant to older clients that don't support the textEdit
property yet. If you have such a client, you should probably support it
anyway, as the legacy insertText doesn't allow the server to communicate
how to replace the prefix that the user already typed. This means that
clients have to guess what to replace, which usually means not replacing
words up to a boundary, not replacing characters such as `$` along with
it.

Fixes #310